Eliminando datos de tablas y creando las que pido cuando ejecuten los seeders

+++ seeders/20241030021803-create-admin-role.js
@@ -2,21 +2,19 @@
 
 module.exports = {
   up: async (queryInterface, Sequelize) => {
-    const [results] = await queryInterface.sequelize.query(
-      "SELECT * FROM Roles WHERE nombreRole = 'Admin'"
-    );
-    if (results.length === 0) {
-      await queryInterface.bulkInsert('Roles', [{
-        nombreRole: 'Admin',
-        createdAt: new Date(),
-        updatedAt: new Date(),
-      }]);
-    }
+    // Borra todos los registros en la tabla Roles
+    await queryInterface.bulkDelete('Roles', null, {});
+
+    // Inserta el rol de Admin
+    await queryInterface.bulkInsert('Roles', [{
+      nombreRole: 'Admin',
+      createdAt: new Date(),
+      updatedAt: new Date(),
+    }], {});
   },
 
   down: async (queryInterface, Sequelize) => {
-    await queryInterface.bulkDelete('Roles', {
-      nombreRole: 'Admin',
-    });
+    // Borra el rol de Admin si es necesario
+    await queryInterface.bulkDelete('Roles', { nombreRole: 'Admin' }, {});
   }
 };
